py/requirements.txt
discord.py==2.3.2
GitPython==3.1.40
pdf2image==1.16.3
Pillow==10.1.0
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.31.0
